Merge branch 'iter' into master
This provides a major update to the API which is significantly more
generic, and generally much more 'rusty'.  This does cause a performance
hit of about 20% over the lower-level API, however I think merging the
new API is justified because:

1: Significantly easier to use and much cleaner code.  It removes a lot
of duplication and boilerplate and allows generating devices that are
format (float vs fixed point) agnostic.

2: Even with the performane hit, I can run 16 voices in floating point
mode using about 20% CPU for the default patch.  Even in fixed-point
mode and with the modmatrix completely maxed out (so the worst case) 16
voices only takes about 50% CPU.

3: I think the new API will likely be more performant on embedded
hardware, which is an explicit goal of the project.

4: I think using SIMD to run multiple voices simultaneously will yield
a pretty significant performance win on x86/floating point.